+++ src/vector.rs
@@ -1,12 +1,12 @@
-pub type Point2D = (f32, f32);
+pub type Point2D = (f64, f64);
 
 pub trait Vector2 {
-    fn new(x: f32, y: f32) -> Self;
+    fn new(x: f64, y: f64) -> Self;
 
-    fn x(self: &Self) -> f32;
-    fn y(self: &Self) -> f32;
+    fn x(self: &Self) -> f64;
+    fn y(self: &Self) -> f64;
 
-    fn gradient(self: &Self, rhs: &Self) -> f32 {
+    fn gradient(self: &Self, rhs: &Self) -> f64 {
         let delta_x = rhs.x() - self.x();
         let delta_y = rhs.y() - self.y();
         delta_y / delta_x
@@ -14,14 +14,14 @@ pub trait Vector2 {
 }
 
 impl Vector2 for Point2D {
-    fn new(x: f32, y: f32) -> Self {
+    fn new(x: f64, y: f64) -> Self {
         (x, y)
     }
 
-    fn x(&self) -> f32 {
+    fn x(&self) -> f64 {
         self.0
     }
-    fn y(&self) -> f32 {
+    fn y(&self) -> f64 {
         self.1
     }
 }
